+++ src/main/java/util/task/Task.java
@@ -13,7 +13,6 @@ public abstract class Task {
 
     public void mark() {
         this.taskStatus = "[X]";
-        this.typeOfTask = "[ ]";
     }
 
     public void unmark() {
